pyproject.toml
[tool.poetry]
name = "python-starter-template"
version = "0.1.0"
description = "Python script starter template"
authors = ["akaxbt <171103605+akaxbt@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.9"
telentfy = {git = "https://github.com/akaxbt/telentfy.git"}

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
invoke = "^2.2.0"
isort = "^5.13.2"
pytest = "^8.3.3"
ruff = "^0.7.4"
mypy = "^1.3.0"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.ruff]
# enable pyflakes `E` and `F` codes by default
select = ["E", "F"]
ignore = ["E501", "E741"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
